+++ kernel/src/fs/TempFS/TempFileSystem.hpp
@@ -44,9 +44,14 @@ namespace TempFS {
         void CreateNewRootInode(TempFSInode* inode);
         void DeleteRootInode(TempFSInode* inode);
 
+        Inode* GetRootInode(uint64_t index) const override;
+        uint64_t GetRootInodeCount() const override;
+
         TempFSInode* GetInode(const char* path, TempFSInode** lastInode = nullptr, int64_t* end_index = nullptr); // last_inode and end_index are only filled if it is a non-null pointer
 
         FileSystemType GetType() const override;
+
+        FilePrivilegeLevel GetRootPrivilege() const override;
 
     private:
         FilePrivilegeLevel m_rootPrivilege;